requirements.txt
Django==1.9
Whoosh==2.7.0
argparse==1.2.1
django-angular==0.7.15
django-appconf==1.0.1
django-filter==0.11
django-haystack==2.4.0
django-statici18n==1.1.5
djangorestframework==3.3.2
six==1.10.0
wsgiref==0.1.2
#cx_freeze==4.3.3
cherrypy
